refactor: KuringNavigator를 domain:navigation 모듈로 옮김
