+++ app/src/main/java/com/ku_stacks/ku_ring/di/NavigatorModule.kt
@@ -1,7 +1,7 @@
 package com.ku_stacks.ku_ring.di
 
+import com.ku_stacks.ku_ring.navigation.KuringNavigator
 import com.ku_stacks.ku_ring.navigator.KuringNavigatorImpl
-import com.ku_stacks.ku_ring.ui_util.KuringNavigator
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
