+++ core/work/src/main/java/com/ku_stacks/ku_ring/work/ReEngagementNotificationWork.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.hilt.work.HiltWorker
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.ku_stacks.ku_ring.ui_util.KuringNavigator
+import com.ku_stacks.ku_ring.navigation.KuringNavigator
 import com.ku_stacks.ku_ring.util.KuringNotificationManager
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
